♻️ Extract Workout to component pattern
